requirements.txt
aniso8601==8.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
